refactor: Extract StatusSettings.bulkAddStatusCollection()
The code was moved from StatusSettingsHelpers.addCustomStatusesCollection(),
and a test was added too.
